Fix the test_executable() implementation.
Modified test_executable() and the backing elf_verify_header to
fall back to FatELF parsing in the case that ELF parsing fails.

Interestingly, part of the code I modified in test_executable seems to
limit the #! line to a maximum size of sizeof(elf_ehdr). That wasn't
immediately relevant to FatELF, so I left that limit in place.