Use CPUID to determine boot-time long-mode support
When running under an ia32 boot loader, the compile time constants can
not be used to determine x86-64 support. Once the kernel is loaded, we
can assume that 64-bit binaries require a 64-bit kernel -- although
that's not strictly necessary:

Mac OS X supports a 64-bit compatible syscall interface and userspace
execution mode from a 32-bit kernel. That was implemented in large part
to support a 32-bit to 64-bit transition, during which time many
3rd-party kernel extensions had not been updated.