Deskbar: Make variable width
resize Deskbar by clicking and dragging the mouse on the
horizontal side of Deskbar opposite the screen's edge

Resize ExpandoMenuBar when window is resized

Rename where to whereScreen to make it clear that the variable
is in screen coordinates.

Lock focus on window while resizing

Refactor

Resize via new TResizeControl class.

Looks and acts like TDragRegion but TDragRegion is more than meets
the eye.

Set default width to minimum.

Don't set width to 0 in settings on quit, we use this setting now.