Deskbar: Fix crash outlined in #13172
Crashed because BarApp->fBarView was NULL at the time we were calling it.

fBarView is set in TBarApp ctor. Apparently not safe to assume it is set even
in AttachedToWindow() or AllAttached(). This never once crashed for me while
testing yet it crashes reliably on hrev50848 nightly, something fishy is going
on.

Fixed by fetching the icon in BarView::AttachedToWindow() instead.